requirements.txt
streamlit
pandas
numpy
tensorflow
scikit-learn
pillow
